% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cobbDouglasPredictions}
\alias{cobbDouglasPredictions}
\title{Cobb-Douglas Predictions}
\usage{
cobbDouglasPredictions(countryAbbrev, energyType, baseHistorical)
}
\value{
a data frame with one variable named \code{pred} containing fitted values on the
natural scale.
}
\description{
Cobb-Douglas Predictions
}

