% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getFolderForResampleData}
\alias{getFolderForResampleData}
\title{Directory for resample data}
\usage{
getFolderForResampleData(modelType = modelTypes,
  countryAbbrev = countryAbbrevs, baseResample)
}
\arguments{
  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev,}{a character string naming the
  country, if you want to use original data.}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data.}
}
\value{
the relative path to the directory containing the data for the requested resample data.
}
\description{
This function returns a string representing the relative directory path containing the resample data being requested.
}

