% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadResampleData}
\alias{loadResampleData}
\title{Loads all resample data, for the base fit and resample fits}
\usage{
loadResampleData(modelType, countryAbbrev, energyType = "none", factor = NA,
  archive = NULL, baseResample)
}
\arguments{
  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev}{a string representing the country
  for which you want to load resample data.}

  \item{energyType}{one of \code{"none"}, \code{"Q"} (for
  thermal energy), \code{"X"} (for exergy), or \code{"U"}
  (for useful work).}

  \item{factor}{one of \code{"K"} (for capital stock),
  \code{"L"} (for labor), \code{"Q"} (for thermal energy),
  \code{"X"} (for exergy), or \code{"U"} (for useful
  work).}

  \item{archive}{the relative path to a .zip archive
  containing resample model data, if resample data are to
  be loaded from the archive}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data, if the data are
  to be loaded from a directory or an archive.}
}
\value{
a data frame containing resample data for specified countries for the given \code{modelType}, \code{energyType},
or \code{factor}.
}
\description{
This function returns a data frame that contains all of the resample model coefficients
}

